zw - user browser history api instead of # history when available
Sets the fallback setting on the dev server historyApiFallback so that
the index will be served when the server does not know the requested
route.

Needed to make the path to index.html and bundle.js absolute so that they
weren't interpreted as realative paths when loading the app with a
url that includes a history path.