zw - adding bank account store and actions that will triggeer changes
The store registers itself as a receiver of events or 'actions' from the
app dispatcher and then handles those events differently based on the type
of the action.  In each case the balance state stored in the store is updated.

The store then fires an event using the emitter.  Those events will be picked up
by various views to trigger re-rendering